api/pyproject.toml
[project]
name = "sso"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.8",
    "fastapi-libkit>=0.1.3",
    "fastapi-sqlalchemy-toolkit>=0.7.11",
    "fastapi-sso>=0.17.0",
    "greenlet>=3.1.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "pyyaml>=6.0.2",
    "asyncpg>=0.30.0",
    "alembic>=1.14.1",
    "minio>=7.2.15",
    "python-magic>=0.4.27",
    "pyjwt>=2.10.1",
    "alembic-postgresql-enum>=1.7.0",
    "crontab>=1.0.1",
]
